split Drive state
