+++ ros/src/computing/planning/decision/packages/decision_maker/CMakeLists.txt
@@ -62,8 +62,9 @@ add_executable(decision_maker_node
         nodes/decision_maker/decision_maker_node_init.cpp
         nodes/decision_maker/decision_maker_node_callback.cpp
         nodes/decision_maker/decision_maker_node_state_vehicle.cpp
-        nodes/decision_maker/decision_maker_node_state_drive.cpp
         nodes/decision_maker/decision_maker_node_state_mission.cpp
+        nodes/decision_maker/decision_maker_node_state_behavior.cpp
+        nodes/decision_maker/decision_maker_node_state_motion.cpp
         nodes/decision_maker/decision_maker_node_publish.cpp
         nodes/decision_maker/cross_road_area.cpp
         )
@@ -91,7 +92,8 @@ install(DIRECTORY launch/
         )
 
 install(FILES
-        autoware_state_drive.yaml
+        autoware_state_behavior.yaml
+        autoware_state_motion.yaml
         autoware_state_mission.yaml
         autoware_state_vehicle.yaml
         DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION})
@@ -108,7 +110,8 @@ if(CATKIN_ENABLE_TESTING)
     nodes/decision_maker/decision_maker_node_init.cpp
     nodes/decision_maker/decision_maker_node_callback.cpp
     nodes/decision_maker/decision_maker_node_state_vehicle.cpp
-    nodes/decision_maker/decision_maker_node_state_drive.cpp
+    nodes/decision_maker/decision_maker_node_state_behavior.cpp
+    nodes/decision_maker/decision_maker_node_state_motion.cpp
     nodes/decision_maker/decision_maker_node_state_mission.cpp
     nodes/decision_maker/decision_maker_node_publish.cpp
     nodes/decision_maker/cross_road_area.cpp)
