+++ ros/src/computing/planning/decision/packages/decision_maker/nodes/decision_maker/decision_maker_node_state_mission.cpp
@@ -134,21 +134,10 @@ void DecisionMakerNode::entryDrivingState(cstring_t& state_name, int status)
 {
   setEventFlag("received_based_lane_waypoint", false);
 
-  if (isEventFlagTrue("emergency_flag"))
-  {
-    tryNextState("mission_aborted");
-    return;
-  }
-
   tryNextState("operation_start");
 }
 void DecisionMakerNode::updateDrivingState(cstring_t& state_name, int status)
 {
-  if (isEventFlagTrue("emergency_flag"))
-  {
-    tryNextState("mission_aborted");
-  }
-
   if (!use_fms_ && auto_mission_change_ && isEventFlagTrue("received_based_lane_waypoint"))
   {
     tryNextState("request_mission_change");
@@ -204,7 +193,11 @@ void DecisionMakerNode::updateMissionChangeFailedState(cstring_t& state_name, in
 void DecisionMakerNode::entryMissionCompleteState(cstring_t& state_name, int status)
 {
   setEventFlag("received_based_lane_waypoint", false);
-  tryNextState("operation_end");
+
+  if (!use_fms_ && auto_mission_reload_)
+    tryNextState("mission_reloaded");
+  else
+    tryNextState("operation_end");
 }
 void DecisionMakerNode::updateMissionCompleteState(cstring_t& state_name, int status)
 {
