+++ ros/src/computing/planning/decision/packages/decision_maker/nodes/decision_maker/decision_maker_node_state_vehicle.cpp
@@ -135,10 +135,6 @@ void DecisionMakerNode::entryVehicleReadyState(cstring_t& state_name, int status
 
 void DecisionMakerNode::updateVehicleReadyState(cstring_t& state_name, int status)
 {
-  if (isEventFlagTrue("emergency_flag"))
-  {
-    tryNextState("emergency");
-  }
 }
 
 void DecisionMakerNode::entryVehicleEmergencyState(cstring_t& state_name, int status)
