+++ ros/src/computing/planning/motion/packages/waypoint_follower/nodes/twist_gate/twist_gate.cpp
@@ -43,7 +43,7 @@ TwistGate::TwistGate(const ros::NodeHandle& nh, const ros::NodeHandle& private_n
   emergency_stop_pub_ = nh_.advertise<std_msgs::Bool>("/emergency_stop", 1, true);
   control_command_pub_ = nh_.advertise<std_msgs::String>("/ctrl_mode", 1);
   vehicle_cmd_pub_ = nh_.advertise<vehicle_cmd_msg_t>("/vehicle_cmd", 1, true);
-  state_cmd_pub_ = nh_.advertise<std_msgs::String>("/state_cmd", 1, true);
+  state_cmd_pub_ = nh_.advertise<std_msgs::String>("/state_cmd", 1, false);
 
   remote_cmd_sub_ = nh_.subscribe("/remote_cmd", 1, &TwistGate::remote_cmd_callback, this);
