+++ ros/src/msgs/autoware_msgs/msg/State.msg
@@ -1,4 +1,5 @@
 Header header
 string vehicle_state
 string mission_state
-string drive_state
+string behavior_state
+string motion_state
